added option to have a custom redisgraph configmap
