+++ controllers/persistentvolumeclaim.go
@@ -1,0 +1,93 @@
+// Copyright (c) 2020 Red Hat, Inc.
+// Copyright Contributors to the Open Cluster Management project
+
+package controllers
+
+import (
+	"context"
+	"time"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+func deletePVC(client client.Client) error {
+	pvc := &corev1.PersistentVolumeClaim{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      pvcName,
+			Namespace: namespace,
+		},
+	}
+	err := client.Delete(context.TODO(), pvc)
+	if err != nil && !errors.IsNotFound(err) {
+		log.Error(err, "Failed to delete search redisgraph PVC", "name", pvcName)
+		return err
+	}
+	time.Sleep(1 * time.Second) //Sleep for a minute to avoid quick update of statefulset
+	log.Info("PVC deleted", "name", pvcName)
+	return nil
+}
+
+func getPVC() *corev1.PersistentVolumeClaim {
+	if storageClass != "" {
+		return &corev1.PersistentVolumeClaim{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      pvcName,
+				Namespace: namespace,
+			},
+			Spec: corev1.PersistentVolumeClaimSpec{
+				AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
+				Resources: corev1.ResourceRequirements{
+					Requests: corev1.ResourceList{
+						corev1.ResourceName(corev1.ResourceStorage): resource.MustParse(storageSize),
+					},
+				},
+				StorageClassName: &storageClass,
+			},
+		}
+	}
+	return &corev1.PersistentVolumeClaim{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      pvcName,
+			Namespace: namespace,
+		},
+		Spec: corev1.PersistentVolumeClaimSpec{
+			AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
+			Resources: corev1.ResourceRequirements{
+				Requests: corev1.ResourceList{
+					corev1.ResourceName(corev1.ResourceStorage): resource.MustParse(storageSize),
+				},
+			},
+		},
+	}
+}
+
+//Remove PVC if you have one
+func setupVolume(client client.Client) error {
+	found := &corev1.PersistentVolumeClaim{}
+	pvc := getPVC()
+	err := client.Get(context.TODO(), types.NamespacedName{Name: pvcName, Namespace: namespace}, found)
+	logKeyPVCName := "PVC Name"
+	if err != nil && errors.IsNotFound(err) {
+		err = client.Create(context.TODO(), pvc)
+		//Return True if sucessfully created pvc else return False
+		if err != nil {
+			log.Info("Error creating a new PVC ", logKeyPVCName, pvcName)
+			log.Info(err.Error())
+			return err
+		} else {
+			log.Info("Created a new PVC ", logKeyPVCName, pvcName)
+			return nil
+		}
+	} else if err != nil {
+		log.Info("Error finding PVC ", logKeyPVCName, pvcName)
+		//return False and error if there is Error
+		return err
+	}
+	log.Info("Using existing PVC")
+	return nil
+}
