+++ controllers/searchoperator_controller.go
@@ -5,23 +5,17 @@ package controllers
 
 import (
 	"context"
-	"crypto/rand"
 	"fmt"
-	"math/big"
 	"os"
 	"reflect"
 	"strconv"
 	"time"
-
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"github.com/go-logr/logr"
 	searchv1alpha1 "github.com/stolostron/search-operator/api/v1alpha1"
 	appv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -398,460 +392,6 @@ func int32Ptr(i int32) *int32 { return &i }
 
 func int64Ptr(i int64) *int64 { return &i }
 
-// compareLabels compares two map[string]string structs
-// Returns false if all key-value pairs in first map is not in second map
-// Else returns true
-// Used to check if all the expected labels are present in the current running statefulset
-func compareLabels(metadataLabels, ssetLabels map[string]string) bool {
-	allLabelsPresent := true
-	for label, value := range metadataLabels {
-		ssetVal, labelPresent := ssetLabels[label]
-		if labelPresent && (ssetVal == value) {
-			continue
-		} else {
-			allLabelsPresent = false
-			log.Info("Not all labels present in statefulset. Label: ", label, "not found in metadata")
-			break
-		}
-	}
-	return allLabelsPresent
-}
-func (r *SearchOperatorReconciler) getStatefulSet(cr *searchv1alpha1.SearchOperator,
-	rdbVolumeSource corev1.VolumeSource, saverdb string) *appv1.StatefulSet {
-	sset := &appv1.StatefulSet{}
-	err := r.Client.Get(context.TODO(), types.NamespacedName{Name: statefulSetName, Namespace: namespace}, sset)
-	if err != nil {
-		r.Log.Info("Error fetching Statefulset")
-	}
-	bool := false
-	metadataLabels := map[string]string{}
-	metadataLabels["release"] = releaseName
-	metadataLabels["component"] = component
-	metadataLabels["app"] = appName
-	if !compareLabels(metadataLabels, sset.Labels) {
-		sset.Labels = metadataLabels
-	}
-	sset.ObjectMeta.Name = statefulSetName
-	sset.ObjectMeta.Namespace = cr.Namespace
-	sset.Spec.Replicas = int32Ptr(1)
-	sset.Spec.Selector = &metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"component": component,
-			"app":       appName,
-		},
-	}
-	sset.Spec.Template.ObjectMeta.Labels = metadataLabels
-	sset.Spec.Template.Spec.ServiceAccountName = "search-operator"
-	tol := corev1.Toleration{
-		Key:      "node-role.kubernetes.io/infra",
-		Effect:   corev1.TaintEffectNoSchedule,
-		Operator: corev1.TolerationOpExists,
-	}
-	sset.Spec.Template.Spec.Tolerations = []corev1.Toleration{tol}
-	pullSecret := corev1.LocalObjectReference{
-		Name: cr.Spec.PullSecret,
-	}
-	sset.Spec.Template.Spec.ImagePullSecrets = []corev1.LocalObjectReference{pullSecret}
-	if sset.Spec.Template.Spec.SecurityContext != nil {
-		sset.Spec.Template.Spec.SecurityContext.FSGroup = int64Ptr(redisUser)
-		sset.Spec.Template.Spec.SecurityContext.RunAsUser = int64Ptr(redisUser)
-	} else {
-		sset.Spec.Template.Spec.SecurityContext = &corev1.PodSecurityContext{
-			FSGroup:   int64Ptr(redisUser),
-			RunAsUser: int64Ptr(redisUser),
-		}
-	}
-	sset.Spec.Template.Spec.Containers = []corev1.Container{
-		{
-			Name:  "redisgraph",
-			Image: cr.Spec.SearchImageOverrides.Redisgraph_TLS,
-			Env: []corev1.EnvVar{
-				{
-					Name: "REDIS_PASSWORD",
-					ValueFrom: &corev1.EnvVarSource{
-						SecretKeyRef: &corev1.SecretKeySelector{
-							LocalObjectReference: corev1.LocalObjectReference{
-								Name: "redisgraph-user-secret",
-							},
-							Key: "redispwd",
-						},
-					},
-				},
-				{
-					Name:  "REDIS_GRAPH_SSL",
-					Value: "true",
-				},
-				{
-					Name:  "SAVERDB",
-					Value: saverdb,
-				},
-			},
-			LivenessProbe: &corev1.Probe{
-				InitialDelaySeconds: 10,
-				TimeoutSeconds:      1,
-				PeriodSeconds:       15,
-				SuccessThreshold:    1,
-				FailureThreshold:    3,
-				ProbeHandler: corev1.ProbeHandler{
-					TCPSocket: &corev1.TCPSocketAction{
-						Port: intstr.FromInt(6380),
-					},
-				},
-			},
-			ReadinessProbe: &corev1.Probe{
-				InitialDelaySeconds: 5,
-				TimeoutSeconds:      1,
-				PeriodSeconds:       15,
-				SuccessThreshold:    1,
-				FailureThreshold:    3,
-				ProbeHandler: corev1.ProbeHandler{
-					TCPSocket: &corev1.TCPSocketAction{
-						Port: intstr.FromInt(6380),
-					},
-				},
-			},
-			Resources: corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					"memory": resource.MustParse(cr.Spec.Redisgraph_Resource.LimitMemory),
-				},
-				Requests: corev1.ResourceList{
-					"cpu":    resource.MustParse(cr.Spec.Redisgraph_Resource.RequestCPU),
-					"memory": resource.MustParse(cr.Spec.Redisgraph_Resource.RequestMemory),
-				},
-			},
-			TerminationMessagePolicy: "File",
-			TerminationMessagePath:   "/dev/termination-log",
-			ImagePullPolicy:          "Always",
-			SecurityContext: &corev1.SecurityContext{
-				Privileged:               &bool,
-				AllowPrivilegeEscalation: &bool,
-			},
-			VolumeMounts: []corev1.VolumeMount{
-				{
-					Name:      "redis-graph-certs",
-					MountPath: "/certs",
-				},
-				{
-					Name:      "stunnel-pid",
-					MountPath: "/rg",
-				},
-			},
-		},
-	}
-	sset.Spec.Template.Spec.Volumes = []corev1.Volume{
-		{
-			Name: "stunnel-pid",
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{},
-			},
-		},
-		{
-			Name: "redis-graph-certs",
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName: "search-redisgraph-certs",
-					Items: []corev1.KeyToPath{
-						{
-							Key:  "tls.crt",
-							Path: "server.crt",
-						},
-						{
-							Key:  "tls.key",
-							Path: "server.key",
-						},
-					},
-					DefaultMode: int32Ptr(420),
-				},
-			},
-		},
-	}
-
-	if (corev1.VolumeSource{}) != rdbVolumeSource {
-		rdbVolume := corev1.Volume{
-			Name:         "persist",
-			VolumeSource: rdbVolumeSource,
-		}
-		sset.Spec.Template.Spec.Volumes = append(sset.Spec.Template.Spec.Volumes, rdbVolume)
-		rdbVolumeMount := corev1.VolumeMount{
-			Name:      "persist",
-			MountPath: "/redis-data",
-		}
-		for i, container := range sset.Spec.Template.Spec.Containers {
-			if container.Name == "redisgraph" {
-				sset.Spec.Template.Spec.Containers[i].VolumeMounts =
-					append(sset.Spec.Template.Spec.Containers[i].VolumeMounts, rdbVolumeMount)
-				log.Info("Added rdbVolumeMount in container: ", container.Name, rdbVolumeMount.MountPath)
-			}
-		}
-	}
-	if cr.Spec.NodeSelector != nil {
-		sset.Spec.Template.Spec.NodeSelector = cr.Spec.NodeSelector
-		log.Info("Added Node Selector")
-	}
-	if err := ctrl.SetControllerReference(cr, sset, r.Scheme); err != nil {
-		log.Info("Cannot set statefulSet OwnerReference", err.Error())
-	}
-	return sset
-}
-func updateCRs(kclient client.Client, operatorCR *searchv1alpha1.SearchOperator, status string,
-	customizationCR *searchv1alpha1.SearchCustomization, persistence bool, storageClass string,
-	storageSize string, customValuesInuse bool) error {
-	var err error
-	err = updateOperatorCR(kclient, operatorCR, status)
-	if err != nil {
-		return err
-	}
-	if customValuesInuse {
-		err = updateCustomizationCR(kclient, customizationCR, persistence, storageClass, storageSize)
-		if err != nil {
-			return err
-		}
-	}
-	log.Info("Updated status in CRs successfully.")
-	return nil
-}
-
-func fetchSrchOperator(kclient client.Client, cr *searchv1alpha1.SearchOperator) (
-	*searchv1alpha1.SearchOperator, error) {
-	found := &searchv1alpha1.SearchOperator{}
-	err := kclient.Get(context.TODO(), types.NamespacedName{Name: cr.Name, Namespace: cr.Namespace}, found)
-	return found, err
-}
-func fetchSrchCustomization(kclient client.Client, cr *searchv1alpha1.SearchCustomization) (
-	*searchv1alpha1.SearchCustomization, error) {
-	found := &searchv1alpha1.SearchCustomization{}
-	err := kclient.Get(context.TODO(), types.NamespacedName{Name: cr.Name, Namespace: cr.Namespace}, found)
-	return found, err
-}
-func updateOperatorCR(kclient client.Client, cr *searchv1alpha1.SearchOperator, status string) error {
-	cr, err := fetchSrchOperator(kclient, cr)
-	if err != nil {
-		log.Error(err, fmt.Sprintf("Failed to get SearchOperator %s/%s ", cr.Namespace, cr.Name))
-		return err
-	}
-	cr.Status.PersistenceStatus = status
-	if deployVarPresent && deployVarErr == nil {
-		cr.Status.DeployRedisgraph = &deploy
-	}
-	err = kclient.Status().Update(context.TODO(), cr)
-	if err != nil {
-		if errors.IsConflict(err) {
-			log.Info("Failed to update status Object has been modified")
-		}
-		log.Info(fmt.Sprintf("Failed to update %s/%s status. Error: %s", cr.Namespace, cr.Name, err.Error()))
-		return err
-	} else {
-		log.Info(fmt.Sprintf("Updated CR status with persistence %s  ", cr.Status.PersistenceStatus))
-	}
-	return nil
-}
-
-func updateCustomizationCR(kclient client.Client, cr *searchv1alpha1.SearchCustomization,
-	persistence bool, storageClass string, storageSize string) error {
-	cr, err := fetchSrchCustomization(kclient, cr)
-	if err != nil {
-		log.Error(err, fmt.Sprintf("Failed to get SearchCustomization %s/%s ", cr.Namespace, cr.Name))
-		return err
-	}
-	cr.Status.Persistence = persistence
-	cr.Status.StorageClass = storageClass
-	cr.Status.StorageSize = storageSize
-	err = kclient.Status().Update(context.TODO(), cr)
-	if err != nil {
-		if errors.IsConflict(err) {
-			log.Info("Failed to update status Object has been modified")
-		}
-		log.Info(fmt.Sprintf("Failed to update %s/%s status. Error:  %s", cr.Namespace, cr.Name, err.Error()))
-		return err
-	} else {
-		log.Info(fmt.Sprintf("Updated CR status with custom persistence %t ", cr.Status.Persistence))
-	}
-	return nil
-}
-
-func statefulSetNeedsUpdate(client client.Client, deployment *appv1.StatefulSet) bool {
-	found := &appv1.StatefulSet{}
-	err := client.Get(context.TODO(), types.NamespacedName{Name: statefulSetName, Namespace: namespace}, found)
-	if err != nil {
-		return true
-	} else {
-		if !reflect.DeepEqual(found.Spec, deployment.Spec) ||
-			!reflect.DeepEqual(found.GetObjectMeta(), deployment.GetObjectMeta()) {
-			log.Info("Volume source and/or metadata needs to be updated for redisgraph Statefulset")
-			return true
-		} else {
-			log.Info("No changes required for Statefulset")
-			return false
-		}
-	}
-}
-func updateRedisStatefulSet(client client.Client, deployment *appv1.StatefulSet) {
-	found := &appv1.StatefulSet{}
-	err := client.Get(context.TODO(), types.NamespacedName{Name: statefulSetName, Namespace: namespace}, found)
-	if err != nil {
-		if errors.IsNotFound(err) {
-			log.Info("Statefulset not found. Creating Statefulset ...")
-			err = client.Create(context.TODO(), deployment)
-			if err != nil {
-				log.Error(err, "Failed to create Statefulset")
-				return
-			}
-			log.Info("Statefulset created successfully")
-			return
-		}
-		log.Error(err, "Failed to fetch Statefulset")
-		return
-	} else {
-		deployment.ObjectMeta.ResourceVersion = found.ObjectMeta.ResourceVersion
-		err = client.Update(context.TODO(), deployment)
-		if err != nil {
-			log.Error(err, "Failed to update Statefulset")
-			return
-		}
-		log.Info("Volume source and/or metadata updated for redisgraph Statefulset")
-		return
-	}
-
-}
-func deleteRedisStatefulSet(client client.Client) error {
-	statefulset := &appv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      statefulSetName,
-			Namespace: namespace,
-		},
-	}
-	err := client.Delete(context.TODO(), statefulset)
-	if err != nil && !errors.IsNotFound(err) {
-		log.Error(err, "Failed to delete search redisgraph statefulset", "name", statefulSetName)
-		return err
-	}
-	time.Sleep(1 * time.Second) //Sleep for a minute to avoid quick update of statefulset
-	log.Info("StatefulSet deleted", "name", statefulSetName)
-	return nil
-}
-
-func deletePVC(client client.Client) error {
-	pvc := &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pvcName,
-			Namespace: namespace,
-		},
-	}
-	err := client.Delete(context.TODO(), pvc)
-	if err != nil && !errors.IsNotFound(err) {
-		log.Error(err, "Failed to delete search redisgraph PVC", "name", pvcName)
-		return err
-	}
-	time.Sleep(1 * time.Second) //Sleep for a minute to avoid quick update of statefulset
-	log.Info("PVC deleted", "name", pvcName)
-	return nil
-}
-
-func getPVC() *corev1.PersistentVolumeClaim {
-	if storageClass != "" {
-		return &corev1.PersistentVolumeClaim{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      pvcName,
-				Namespace: namespace,
-			},
-			Spec: corev1.PersistentVolumeClaimSpec{
-				AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-				Resources: corev1.ResourceRequirements{
-					Requests: corev1.ResourceList{
-						corev1.ResourceName(corev1.ResourceStorage): resource.MustParse(storageSize),
-					},
-				},
-				StorageClassName: &storageClass,
-			},
-		}
-	}
-	return &corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pvcName,
-			Namespace: namespace,
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-			Resources: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceName(corev1.ResourceStorage): resource.MustParse(storageSize),
-				},
-			},
-		},
-	}
-}
-
-//Remove PVC if you have one
-func setupVolume(client client.Client) error {
-	found := &corev1.PersistentVolumeClaim{}
-	pvc := getPVC()
-	err := client.Get(context.TODO(), types.NamespacedName{Name: pvcName, Namespace: namespace}, found)
-	logKeyPVCName := "PVC Name"
-	if err != nil && errors.IsNotFound(err) {
-		err = client.Create(context.TODO(), pvc)
-		//Return True if sucessfully created pvc else return False
-		if err != nil {
-			log.Info("Error creating a new PVC ", logKeyPVCName, pvcName)
-			log.Info(err.Error())
-			return err
-		} else {
-			log.Info("Created a new PVC ", logKeyPVCName, pvcName)
-			return nil
-		}
-	} else if err != nil {
-		log.Info("Error finding PVC ", logKeyPVCName, pvcName)
-		//return False and error if there is Error
-		return err
-	}
-	log.Info("Using existing PVC")
-	return nil
-}
-
-func generatePass(length int) []byte {
-	chars := "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
-		"abcdefghijklmnopqrstuvwxyz" +
-		"0123456789"
-
-	buf := make([]byte, length)
-	for i := 0; i < length; i++ {
-		nBig, _ := rand.Int(rand.Reader, big.NewInt(int64(len(chars))))
-		buf[i] = chars[nBig.Int64()]
-	}
-	return buf
-}
-
-// newRedisSecret returns a redisgraph-user-secret with the same name/namespace as the cr
-func newRedisSecret(cr *searchv1alpha1.SearchOperator, scheme *runtime.Scheme) *corev1.Secret {
-	labels := map[string]string{
-		"app": "search",
-	}
-
-	sec := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "redisgraph-user-secret",
-			Namespace: namespace,
-			Labels:    labels,
-		},
-		Data: map[string][]byte{
-			"redispwd": generatePass(16),
-		},
-	}
-	if err := ctrl.SetControllerReference(cr, sec, scheme); err != nil {
-		log.Info("Cannot set secret OwnerReference", err.Error())
-	}
-	return sec
-}
-
-func isStatefulSetAvailable(kclient client.Client) bool {
-	//check if statefulset is present if not we can assume the pod is not running
-	found := &appv1.StatefulSet{}
-	err := kclient.Get(context.TODO(), types.NamespacedName{Name: statefulSetName, Namespace: namespace}, found)
-	if err != nil && errors.IsNotFound(err) {
-		return false
-	}
-	return true
-}
-
 func (r *SearchOperatorReconciler) isPodRunning(withPVC bool, waitSeconds int) bool {
 	log.Info("Checking Redisgraph Pod Status...")
 	//Keep checking status until waitSeconds
@@ -972,15 +512,4 @@ func (r *SearchOperatorReconciler) setupSecret(client client.Client, cr *searchv
 		log.Info("Skip reconcile: Secret already exists", "Secret.Namespace", found.Namespace, "Secret.Name", found.Name)
 	}
 	return nil
-}
-
-func getOptions(opts map[string]string) []client.ListOption {
-	listOptions := []client.ListOption{}
-	listOption := client.ListOptions{
-		LabelSelector: labels.SelectorFromSet(opts),
-		Namespace:     namespace,
-		Limit:         2, //setting this to 2 as a safety net while deleting pods
-	}
-	listOptions = append(listOptions, &listOption)
-	return listOptions
 }
